moved _get_regridding_domain
